requirements.txt
asgiref==3.7.2
distlib==0.3.7
dj-database-url==2.1.0
Django==4.2.7
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.14.0
filelock==3.12.4
install==1.3.5
numpy==1.26.1
platformdirs==3.11.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-decouple==3.8
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
virtualenv==20.24.5
